Move DummyAWG to tests
